renames action to response body
