+++ src/main/java/de/rabea/response/body/EmptyResponse.java
@@ -1,0 +1,11 @@
+package de.rabea.response.body;
+
+import de.rabea.server.ResponseBody;
+
+public class EmptyResponse implements ResponseBody {
+
+    @Override
+    public byte[] response() {
+        return "".getBytes();
+    }
+}
