--- src/main/java/de/rabea/server/action/EmptyResponse.java
@@ -1,11 +1,0 @@
-package de.rabea.server.action;
-
-import de.rabea.server.Action;
-
-public class EmptyResponse implements Action {
-
-    @Override
-    public byte[] response() {
-        return "".getBytes();
-    }
-}
